Moved Player to Interactions package, added "exit" word to Game
