+++ src/main/java/org/vashonsd/pirateship/commands/Command.java
@@ -3,8 +3,8 @@ package org.vashonsd.pirateship.commands;
 
 import java.util.ArrayList;
 
-import org.vashonsd.pirateship.Player;
 import org.vashonsd.pirateship.interactions.Actor;
+import org.vashonsd.pirateship.interactions.Player;
 import org.vashonsd.pirateship.interactions.Response;
 import org.vashonsd.pirateship.interactions.VisibilityLevel;
