+++ src/main/java/org/vashonsd/pirateship/commands/Take.java
@@ -1,7 +1,7 @@
 package org.vashonsd.pirateship.commands;
 
-import org.vashonsd.pirateship.Player;
 import org.vashonsd.pirateship.interactions.Actor;
+import org.vashonsd.pirateship.interactions.Player;
 import org.vashonsd.pirateship.interactions.Response;
 
 public class Take extends Command {
