+++ src/test/java/org/vashonsd/pirateship/interactions/Leap.java
@@ -1,6 +1,5 @@
 package org.vashonsd.pirateship.interactions;
 
-import org.vashonsd.pirateship.Player;
 import org.vashonsd.pirateship.commands.EasterEgg;
 import org.vashonsd.pirateship.interactions.Actor;
 import org.vashonsd.pirateship.interactions.Response;
